잔잔바리 수정
